requirements.txt
uvicorn~=0.32.1
fastapi~=0.115.6
pydantic~=2.10.3
joblib~=1.4.2
numpy~=2.2.6
pydub~=0.25.1
requests~=2.31.0
python-dotenv~=1.0.1